issue #23 (фильтрация по статусу) issue #18 (вылет приложения после удаления людей, без удаления лис) убран статус "дома", так как регистраторы его не отслеживают

+++ app/src/main/java/ru/kireev/mir/registrarlizaalert/ui/main/AllVolunteersFragment.kt
@@ -6,6 +6,7 @@ import android.net.Uri
 import android.os.Bundle
 import android.view.*
 import android.widget.SearchView
+import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
@@ -13,12 +14,14 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.fragment_tabbed_all_volunteers.view.*
 import ru.kireev.mir.registrarlizaalert.R
 import ru.kireev.mir.registrarlizaalert.adapters.VolunteerAdapter
+import ru.kireev.mir.registrarlizaalert.data.FoxesViewModel
 import ru.kireev.mir.registrarlizaalert.data.Volunteer
 import ru.kireev.mir.registrarlizaalert.data.VolunteersViewModel
 import ru.kireev.mir.registrarlizaalert.listeners.OnVolunteerPhoneNumberClickListener
 
 class AllVolunteersFragment : Fragment(), View.OnClickListener, SearchView.OnQueryTextListener {
 
+    private lateinit var foxesViewModel: FoxesViewModel
     private lateinit var viewModel: VolunteersViewModel
     private lateinit var adapter: VolunteerAdapter
     private var fullList = listOf<Volunteer>()
@@ -26,6 +29,9 @@ class AllVolunteersFragment : Fragment(), View.OnClickListener, SearchView.OnQue
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         val root = inflater.inflate(R.layout.fragment_tabbed_all_volunteers, container, false)
         val recyclerView = root.recyclerViewAllVolunteersTab
+        val foxModel by viewModels<FoxesViewModel>()
+        foxesViewModel = foxModel
+        foxesViewModel.allFoxes.observe(viewLifecycleOwner, Observer {  })
         val model by viewModels<VolunteersViewModel>()
         viewModel = model
         adapter = VolunteerAdapter(requireContext(), model)
@@ -53,7 +59,12 @@ class AllVolunteersFragment : Fragment(), View.OnClickListener, SearchView.OnQue
         alertDialog.setTitle(getString(R.string.warning))
         alertDialog.setMessage(getString(R.string.message_confirm_delete_all))
         alertDialog.setPositiveButton(getString(R.string.delete_all)) { _, _ ->
-            viewModel.deleteAllVolunteers()
+            if (foxesViewModel.allFoxes.value.isNullOrEmpty()) {
+                viewModel.deleteAllVolunteers()
+            } else {
+                Toast.makeText(context, "Перед удалением волонтеров необходимо удалить все лисы!", Toast.LENGTH_SHORT).show()
+            }
+
         }
         alertDialog.setNegativeButton(getString(R.string.cancel), null)
         alertDialog.show()
